+++ pythonFiles/include/blender_vscode/environment.py
@@ -3,6 +3,7 @@ import os
 import sys
 from pathlib import Path
 from typing import Optional
+from typing import Tuple
 
 import addon_utils
 import bpy
@@ -17,13 +18,14 @@ _str_to_log_level = {
 }
 
 
-def _parse_log(env_var_name: str) -> int:
+def _parse_log(env_var_name: str) -> Tuple[int, bool]:
     log_env_global = os.environ.get(env_var_name, "info") or "info"
     try:
-        return _str_to_log_level[log_env_global]
+        enable_flask_logs = "debug-with-flask" == log_env_global
+        return _str_to_log_level[log_env_global], enable_flask_logs
     except KeyError as e:
         logging.warning(f"Log level for {env_var_name} not set: {e}")
-        return logging.WARNING
+        return logging.WARNING, False
 
 
 # binary_path_python was removed in blender 2.92
@@ -38,5 +40,6 @@ scripts_folder = blender_path.parent / f"{version[0]}.{version[1]}" / "scripts"
 addon_directories = tuple(map(Path, addon_utils.paths()))
 
 EXTENSIONS_REPOSITORY: Optional[str] = os.environ.get("VSCODE_EXTENSIONS_REPOSITORY", "user_default") or "user_default"
-LOG_LEVEL_GLOBAL = _parse_log("VSCODE_GLOBAL_LOG_LEVEL")
-LOG_LEVEL = _parse_log("VSCODE_LOG_LEVEL")
+LOG_LEVEL, LOG_FLASK = _parse_log("VSCODE_LOG_LEVEL")
+
+logging.getLogger("werkzeug").setLevel(logging.DEBUG if LOG_FLASK else logging.ERROR)