+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- You can now configure VS code internal log level using [`blender.addon.logLevel`](vscode://settings/blender.addon.logLevel) and [`blender.addon.logLevelGlobal`](vscode://settings/blender.addon.logLevelGlobal) ([#198](https://github.com/JacquesLucke/blender_vscode/pull/198))
+- You can now configure VS code internal log level using [`blender.addon.logLevel`](vscode://settings/blender.addon.logLevel) ([#198](https://github.com/JacquesLucke/blender_vscode/pull/198))
   - to mute logs set log level to critical
   - to enable more logs set log level to debug
   - changing log level required blender restart